Command line usage has been changed: stoptime is the first parameter, filename is optional since the program writes on the standard output

+++ src/testSM.py
@@ -1,36 +1,38 @@
-#!/usr/bin/python3
 import simpy
 import sys
 import os
 import multiprocessing
-from pydispatch import dispatcher
 from logging import Logger
 from gate import Gate
 
-def makeLogger(l,q):
-	logger = Logger(l,q)
+def makeLogger(q):
+	logger = Logger(q)
 	logger.run()
 
-def makeLogging(logname):
+def makeLogging():
 	queue = multiprocessing.Queue()
-	p = multiprocessing.Process(target=makeLogger, args=(logname,queue,))
+	p = multiprocessing.Process(target=makeLogger, args=(queue,))
 	p.start()
 	return queue
 		
-def main(fname,stop):
+def main(stop):
 	env = simpy.Environment()
-	sendqueue = makeLogging(fname)
+	sendqueue = makeLogging()
 	Gate(env,sendqueue,'GA',0.05,100)
 	env.run(until=stop)
 	sendqueue.put('HALT')
 
 if __name__ == "__main__":
-	if len(sys.argv) > 2:
-		fname = sys.argv[1]
-		stopTime = int(sys.argv[2])
-		if os.path.exists(fname):
-			os.remove(fname)
-		main(fname,stopTime)
+	if len(sys.argv) > 1:
+		stopTime = int(sys.argv[1])
+		original = sys.stdout
+		if len(sys.argv) > 2:
+			fname = sys.argv[2]
+			if os.path.exists(fname):
+				os.remove(fname)
+			original = open(fname,"w")
+		main(stopTime)
+		sys.stdout = original
 	else:
-		usage = sys.argv[0] + ' logname stoptime' 
+		usage = sys.argv[0] + ' stoptime logfilename'
 		print(usage)
